test: reorganize test suites
