+++ grobl-cli/tests/integration/__init__.py
@@ -1,0 +1,1 @@
+"""Integration tests for the grobl CLI project."""
