+++ grobl-cli/tests/unit/test_unit_prompt.py
@@ -1,3 +1,7 @@
+"""Unit tests for grobl_cli.service.prompt utilities."""
+
+from __future__ import annotations
+
 from pathlib import Path
 from unittest import mock
 
@@ -6,17 +10,17 @@ import pytest
 from grobl_cli.service import prompt
 
 
-def test_env_assume_yes_true(monkeypatch):
+def test_env_assume_yes_true(monkeypatch: pytest.MonkeyPatch) -> None:
     monkeypatch.setenv("GROBL_ASSUME_YES", "true")
     assert prompt.env_assume_yes() is True
 
 
-def test_env_assume_yes_false(monkeypatch):
+def test_env_assume_yes_false(monkeypatch: pytest.MonkeyPatch) -> None:
     monkeypatch.setenv("GROBL_ASSUME_YES", "no")
     assert prompt.env_assume_yes() is False
 
 
-def test_warn_skipped_if_assume_yes(monkeypatch):
+def test_warn_skipped_if_assume_yes(monkeypatch: pytest.MonkeyPatch) -> None:
     # Should not call confirm if assume_yes is True
     paths = (Path(),)
     with mock.patch("grobl_cli.service.prompt._detect_heavy_dirs", return_value={"node_modules"}):
@@ -28,7 +32,7 @@ def test_warn_skipped_if_assume_yes(monkeypatch):
         )
 
 
-def test_warn_shown_and_aborted(monkeypatch):
+def test_warn_shown_and_aborted(monkeypatch: pytest.MonkeyPatch) -> None:
     paths = (Path(),)
     with mock.patch("grobl_cli.service.prompt._detect_heavy_dirs", return_value={"node_modules"}):
         with pytest.raises(SystemExit) as excinfo:
@@ -41,7 +45,7 @@ def test_warn_shown_and_aborted(monkeypatch):
         assert excinfo.value.code == 2
 
 
-def test_warn_shown_and_accepted(monkeypatch):
+def test_warn_shown_and_accepted(monkeypatch: pytest.MonkeyPatch) -> None:
     paths = (Path(),)
     with mock.patch("grobl_cli.service.prompt._detect_heavy_dirs", return_value={"node_modules"}):
         # Should not raise
