+++ grobl-config/tests/system/__init__.py
@@ -1,0 +1,1 @@
+"""System tests for the grobl-config package."""
