--- grobl/tests/test_grobl.py
@@ -1,5 +1,0 @@
-import grobl
-
-
-def test_import():
-    assert grobl
